IMP - Refactor file structure
