+++ src/components/elements/gameData/skillAtk/in/sections/atk.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 
-import {useI18n} from '../../../../../i18n/hook';
-import {NumericInput} from '../../../common/input/numeric';
-import {SectionTitle} from '../../elements/title';
-import {SectionProps} from './props';
+import {useI18n} from '../../../../../../i18n/hook';
+import {NumericInput} from '../../../../common/input/numeric';
+import {SectionTitle} from '../../../elements/title';
+import {SectionProps} from '../props';
+
 
 export const SectionAtk = ({inputData, setInputData}: SectionProps) => {
   const {t} = useI18n();
