+++ src/components/elements/gameData/skillAtk/in/sections/crt.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
 
-import {useI18n} from '../../../../../i18n/hook';
-import {InlineCheck} from '../../../common/check/inlineCheck';
-import {NumericInput} from '../../../common/input/numeric';
-import {SectionTitle} from '../../elements/title';
-import {SectionProps} from './props';
+import {useI18n} from '../../../../../../i18n/hook';
+import {InlineCheck} from '../../../../common/check/inlineCheck';
+import {NumericInput} from '../../../../common/input/numeric';
+import {SectionTitle} from '../../../elements/title';
+import {SectionProps} from '../props';
+
 
 export const SectionCrt = ({inputData, setInputData}: SectionProps) => {
   const {t} = useI18n();
