+++ src/components/elements/gameData/skillAtk/in/sections/filter.tsx
@@ -2,16 +2,17 @@ import React from 'react';
 
 import Form from 'react-bootstrap/Form';
 
-import {useI18n} from '../../../../../i18n/hook';
-import {EnumChecksBox} from '../../../common/check/enumChecksBox';
-import {InlineCheck} from '../../../common/check/inlineCheck';
-import {SectionSubTitle} from '../../elements/subTitle';
-import {SectionTitle} from '../../elements/title';
+import {useI18n} from '../../../../../../i18n/hook';
+import {EnumChecksBox} from '../../../../common/check/enumChecksBox';
+import {InlineCheck} from '../../../../common/check/inlineCheck';
+import {SectionSubTitle} from '../../../elements/subTitle';
+import {SectionTitle} from '../../../elements/title';
 import {
   SectionProps,
   SectionPropsCondEnums,
   SectionPropsElemEnums,
-} from './props';
+} from '../props';
+
 
 type SectionFilterProps =
   SectionProps &
